Move progressbar and console into same page
