--- rhinosystem/gtk/console.ui
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<interface>
-  <requires lib="gtk" version="4.0"/>
-  <requires lib="Adw" version="1.0"/>
-  <template class="ConsoleView" parent="GtkBox">
-    <property name="orientation">vertical</property>
-    <property name="valign">fill</property>
-    <property name="halign">fill</property>
-
-    <child>
-      <object class="GtkBox">
-        <property name="margin-start">40px</property>
-        <property name="margin-end">40px</property>
-        <property name="margin-top">20px</property>
-        <property name="valign">fill</property>
-        <property name="halign">fill</property>
-        <property name="vexpand">true</property>
-        <property name="hexpand">true</property>
-        <child>
-          <object class="GtkBox" id="log_box">
-            <property name="halign">fill</property>
-            <property name="valign">fill</property>
-          </object>
-        </child>
-        <style>
-          <class name="card"/>
-        </style>
-      </object>
-    </child>
-    <child>
-      <object class="GtkButton" id="log_button">
-       <property name="halign">center</property>
-        <property name="valign">center</property>
-        <property name="margin-top">20px</property>
-        <property name="margin-bottom">20px</property>
-        <property name="label">Show terminal output</property>
-        <style>
-          <class name="pill"/>
-        </style>
-      </object>
-    </child>
-  </template>
-</interface>
