+++ rhinosystem/gtk/upgrade.ui
@@ -5,12 +5,51 @@
   <template class="UpgradeView" parent="GtkBox">
     <property name="orientation">vertical</property>
     <child>
-      <object class="GtkStack" id="view_stack">
-        <property name="transition-type">crossfade</property>
+      <object class="GtkProgressBar" id="progress_bar">
+        <property name="visible">true</property>
+        <property name="margin-start">40px</property>
+        <property name="margin-end">40px</property>
+        <property name="valign">center</property>
+        <property name="halign">fill</property>
         <property name="vexpand">true</property>
         <property name="hexpand">true</property>
+        <property name="show-text">true</property>
+        <property name="text">Downloading updates</property>
+        <style>
+          <class name="progress"/>
+        </style>
+      </object>
+    </child>
+    <child>
+      <object class="GtkBox">
+        <property name="margin-start">40px</property>
+        <property name="margin-end">40px</property>
+        <property name="margin-top">20px</property>
         <property name="valign">fill</property>
         <property name="halign">fill</property>
+        <property name="vexpand">true</property>
+        <property name="hexpand">true</property>
+        <child>
+          <object class="GtkBox" id="log_box">
+            <property name="halign">fill</property>
+            <property name="valign">fill</property>
+          </object>
+        </child>
+        <style>
+          <class name="card"/>
+        </style>
+      </object>
+    </child>
+    <child>
+      <object class="GtkButton" id="show_log">
+        <property name="halign">center</property>
+        <property name="valign">center</property>
+        <property name="margin-top">20px</property>
+        <property name="margin-bottom">20px</property>
+        <property name="label">Show terminal output</property>
+        <style>
+          <class name="pill"/>
+        </style>
       </object>
     </child>
   </template>
