+++ rhinosystem/rhinosystem.gresource.xml
@@ -5,8 +5,6 @@
     <file preprocess="xml-stripblanks">gtk/help-overlay.ui</file>
     <file preprocess="xml-stripblanks">gtk/sysinfo.ui</file>
     <file preprocess="xml-stripblanks">gtk/upgrade.ui</file>
-    <file preprocess="xml-stripblanks">gtk/progressbar.ui</file>
-    <file preprocess="xml-stripblanks">gtk/console.ui</file>
     <file preprocess="xml-stripblanks">widgets/inforow.ui</file>
     <file>style.css</file>
   </gresource>
