+++ rhinosystem/utils/deviceinfo.py
