+++ rhinosystem/views/upgrade.py
@@ -19,42 +19,43 @@
 from gi.repository import Adw
 from gi.repository import Gtk
 from gi.repository import GLib
+from gi.repository import Vte
+from gi.repository import Pango
 import time
 from rhinosystem.utils.threading import RunAsync
-from rhinosystem.views.upgradepages.progressbar import ProgressView
-from rhinosystem.views.upgradepages.console import ConsoleView
 
 @Gtk.Template(resource_path='/org/rhinolinux/system/gtk/upgrade.ui')
 class UpgradeView(Gtk.Box):
     __gtype_name__ = "UpgradeView"
 
-    view_stack: Gtk.Stack = Gtk.Template.Child()
+    log_box: Gtk.Box = Gtk.Template.Child()
+    progress_bar: Gtk.ProgressBar = Gtk.Template.Child()
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.log=False
         self.window = None
-        self.progress = ProgressView()
-        self.console = ConsoleView()
-        self.progress.show_log.connect("clicked", self.do_show_log)
-        self.console.log_button.connect("clicked", self.do_show_log)
-        self.view_stack.add_child(self.progress)
-        self.view_stack.add_child(self.console)
-        self.view_stack.set_visible_child(self.progress)
+        self.vte_instance = Vte.Terminal()
+        self.vte_instance.set_cursor_blink_mode(Vte.CursorBlinkMode.ON)
+        self.vte_instance.set_mouse_autohide(True)
+        self.vte_instance.set_font(Pango.FontDescription("Source Code Pro Regular 12"))
+        self.log_box.append(self.vte_instance)
 
+    def on_show(self):
+        self.vte_instance.spawn_async(
+            Vte.PtyFlags.DEFAULT,
+            ".",  # working directory
+            ["rpk", "update", "-y"],
+            [],  # environment
+            GLib.SpawnFlags.DO_NOT_REAP_CHILD,
+            None,
+            None,
+            -1,
+            None,
+            None,
+        )
 
     def on_show(self, window):
         self.window = window
-        self.progress.progress_bar.set_show_text(True)
-        self.progress.progress_bar.set_text("Upgrading Rhino Linux")
-
-        self.console.on_show()
-
-
-    def do_show_log(self, widget):
-        if self.log:
-            self.view_stack.set_visible_child(self.progress)
-            self.log = False
-        else:
-            self.view_stack.set_visible_child(self.console)
-            self.log = True
+        self.progress_bar.set_show_text(True)
+        self.progress_bar.set_text("Upgrading Rhino Linux")
